+++ src/pages/PullRequestPage/Summary/hooks.spec.js
@@ -9,7 +9,7 @@ jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'), // import and retain the original functionalities
   useParams: jest.fn(() => {}),
 }))
-jest.mock('services/pull/hooks')
+jest.mock('services/pull')
 
 const pull = {
   pullId: 5,
