+++ src/services/navigation/index.js
@@ -1,3 +1,3 @@
-export * from './hooks'
+export * from './useLocationParams'
 export * from './normalize'
 export * from './useNavLinks'
