+++ src/services/navigation/useLocationParams.spec.js
@@ -2,7 +2,7 @@ import { renderHook } from '@testing-library/react-hooks'
 import { createMemoryHistory } from 'history'
 import { Route, Router } from 'react-router-dom'
 
-import { useLocationParams } from './hooks'
+import { useLocationParams } from './useLocationParams'
 
 describe('useLocationParams', () => {
   let hookData
