+++ src/services/orgUploadToken/index.js
@@ -1,1 +1,1 @@
-export * from './hooks'
+export * from './useRegenerateOrgUploadToken'
