+++ src/services/orgUploadToken/useRegenerateOrgUploadToken.spec.js
@@ -4,12 +4,12 @@ import { graphql } from 'msw'
 import { setupServer } from 'msw/node'
 import { MemoryRouter, Route } from 'react-router-dom'
 
-import { useRegenerateProfilingToken } from './hooks'
+import { useRegenerateOrgUploadToken } from './useRegenerateOrgUploadToken'
 
 const data = {
   data: {
-    regenerateProfilingToken: {
-      profilingToken: 'new token',
+    regenerateOrgUploadToken: {
+      orgUploadToken: 'new token',
     },
   },
 }
@@ -29,16 +29,16 @@ const wrapper = ({ children }) => (
   </MemoryRouter>
 )
 
-describe('useRegenerateProfilingToken', () => {
+describe('useRegenerateOrgUploadToken', () => {
   let hookData
 
   function setup() {
     server.use(
-      graphql.mutation('regenerateProfilingToken', (req, res, ctx) => {
+      graphql.mutation('regenerateOrgUploadToken', (req, res, ctx) => {
         return res(ctx.status(200), ctx.data({ data }))
       })
     )
-    hookData = renderHook(() => useRegenerateProfilingToken(), {
+    hookData = renderHook(() => useRegenerateOrgUploadToken(), {
       wrapper,
     })
   }
