+++ src/services/profilingToken/index.js
@@ -1,1 +1,1 @@
-export * from './hooks'
+export * from './useRegenerateProfilingToken'
