+++ src/services/pulls/index.js
@@ -1,1 +1,1 @@
-export * from './hooks'
+export * from './usePulls'
